requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2020.12.5
    # via httpx
cffi==1.14.5
    # via cryptography
click==7.1.2
    # via uvicorn
cryptography==3.4.7
    # via python-jose
ecdsa==0.14.1
    # via python-jose
fastapi==0.63.0
    # via -r requirements.in
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.12.3
    # via httpx
httptools==0.1.1
    # via uvicorn
httpx==0.17.1
    # via -r requirements.in
idna==3.1
    # via rfc3986
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.20
    # via cffi
pydantic==1.8.1
    # via fastapi
python-dotenv==0.16.0
    # via uvicorn
python-jose[cryptography]==3.2.0
    # via -r requirements.in
pyyaml==5.4.1
    # via uvicorn
rfc3986[idna2008]==1.4.0
    # via httpx
rsa==4.7.2
    # via python-jose
six==1.15.0
    # via
    #   ecdsa
    #   python-jose
sniffio==1.2.0
    # via
    #   httpcore
    #   httpx
starlette==0.13.6
    # via fastapi
typing-extensions==3.7.4.3
    # via pydantic
uvicorn[standard]==0.13.4
    # via -r requirements.in
uvloop==0.15.2
    # via uvicorn
watchgod==0.7
    # via uvicorn
websockets==8.1
    # via uvicorn
